Add files via upload
the Google Play Store team is working on a feature that will allow them to highlight the most promising apps. This analysis would now necessitate a basic understanding of the characteristics that make a high-performing app. You might pose questions like the following:

Does an app being larger or more expensive inevitably indicate that it will perform better than the others?
Is it true that a bigger number of instals indicates which app will get a higher rating than others?